Adding factory for qrsInteract.
