+++ qrsInteract.js
@@ -1,78 +1,12 @@
 var config = require('./config/config');
 var common = require('./common');
-var request = require('request');
 var fs = require('fs');
 var Promise = require('bluebird');
 var winston = require('winston');
 var extend = require('extend');
 var path = require('path');
 
-var qrsInteract = function QRSInteract(inputConfig) {
-    var generateXrfKey = function() {
-        var xrfString = "";
-        for (i = 0; i < 16; i++) {
-            if (Math.floor(Math.random() * 2) == 0) {
-                xrfString += Math.floor(Math.random() * 10).toString();
-            } else {
-                var charNumber = Math.floor(Math.random() * 26);
-                if (Math.floor(Math.random() * 2) == 0) {
-                    //small letter
-                    xrfString += String.fromCharCode(charNumber + 97);
-                } else {
-                    xrfString += String.fromCharCode(charNumber + 65);
-                }
-            }
-        }
-        return xrfString;
-    }
-
-    var updateConfig = function(inputConfig)
-    {
-        var newConfig = common.clone(config);
-        if (typeof inputConfig == 'string')
-        {
-            newConfig.hostname = inputConfig;
-        }
-        else
-        {
-            newConfig = extend(true, newConfig, inputConfig);
-        }
-        
-        newConfig = extend(true, newConfig, {
-            certificates: {
-		        client: path.resolve(newConfig.localCertPath, 'client.pem'),
-		        client_key: path.resolve(newConfig.localCertPath, 'client_key.pem'),
-		        server: path.resolve(newConfig.localCertPath, 'server.pem'),
-		        server_key: path.resolve(newConfig.localCertPath, 'server_key.pem'),
-		        root: path.resolve(newConfig.localCertPath, 'root.pem')
-            }
-        });
-
-        return newConfig;
-    }
-
-    var localConfig = updateConfig(inputConfig);
-
-
-    var defaultPort = localConfig.portNumber;
-    var basePath = "https://" + localConfig.hostname + ":" + defaultPort + "/qrs";
-    var xrfkey = generateXrfKey();
-    var xrfkeyParam = "xrfkey="+xrfkey;
-    var defaults = request.defaults({
-        rejectUnauthorized: false,
-        host: localConfig.hostname,
-        cert: fs.readFileSync(localConfig.certificates.client),
-        key: fs.readFileSync(localConfig.certificates.client_key),
-        ca: fs.readFileSync(localConfig.certificates.root),
-        headers: {
-            'X-Qlik-User': localConfig.repoAccount,
-            'X-Qlik-Xrfkey': xrfkey,
-            'Content-Type': 'application/json',
-            'Accept-Encoding': 'gzip'
-        },
-        gzip: true,
-        json: true
-    });
+var qrsInteract = function QRSInteractMain(basePath, xrfkeyParam, requestDefaults) {
 
     if (!String.prototype.startsWith) {
         String.prototype.startsWith = function(searchString, position) {
@@ -91,23 +25,6 @@ var qrsInteract = function QRSInteract(inputConfig) {
             var lastIndex = subjectString.indexOf(searchString, position);
             return lastIndex !== -1 && lastIndex === position;
         };
-    }
-
-    var generateXrfKey = function() {
-        var xrfString = "";
-        for (i = 0; i < 16; i++) {
-            if (Math.floor(Math.random() * 2) == 0) {
-                xrfString += Math.floor(Math.random() * 10).toString();
-            } else {
-                var charNumber = Math.floor(Math.random() * 26);
-                if (Math.floor(Math.random() * 2) == 0) {
-                    //small letter
-                    xrfString += String.fromCharCode(charNumber + 97);
-                } else {
-                    xrfString += String.fromCharCode(charNumber + 65);
-                }
-            }
-        }
     }
 
     var getFullPath = function(path) {
@@ -135,7 +52,7 @@ var qrsInteract = function QRSInteract(inputConfig) {
         return new Promise(function(resolve, reject) {
             path = getFullPath(path);
             var sCode;
-            var r = defaults;
+            var r = requestDefaults;
             var res = '';
             r.get(path)
                 .on('response', function(response, body) {
@@ -163,7 +80,7 @@ var qrsInteract = function QRSInteract(inputConfig) {
         return new Promise(function(resolve, reject) {
             path = getFullPath(path);
             var sCode;
-            var r = defaults;
+            var r = requestDefaults;
             var res = '';
             var finalBody = body != undefined ? (sendType.toLowerCase() == 'json' ? body : JSON.stringify(body)) : undefined;
             r({
@@ -194,7 +111,7 @@ var qrsInteract = function QRSInteract(inputConfig) {
         return new Promise(function(resolve, reject) {
             path = getFullPath(path);
             var sCode;
-            var r = defaults;
+            var r = requestDefaults;
             r({
                     url: path,
                     method: 'PUT',
@@ -218,7 +135,7 @@ var qrsInteract = function QRSInteract(inputConfig) {
         return new Promise(function(resolve, reject) {
             path = getFullPath(path);
             var sCode;
-            var r = defaults;
+            var r = requestDefaults;
 
             r({
                     url: path,
@@ -241,7 +158,7 @@ var qrsInteract = function QRSInteract(inputConfig) {
 
     this.GetBasePath = function()
     {
-        return "https://" + localConfig.hostname + ":" + defaultPort + "/qrs";
+        return basePath;
     }
 };
